put models in seperate directory
