+++ src/models/duplicate-file-finder.ts
@@ -1,7 +1,13 @@
 import fs from 'fs';
 import fsPromise from 'fs/promises';
-import { IFileInfo, config, IMapedFile } from './find-duplicate-files';
+import { IFileInfo } from '..';
+import { defaultConfig } from '../default-config';
 import { Util } from './util';
+
+interface IMapedFile {
+  name: string;
+  buffer: Buffer;
+}
 
 export interface IDuplicateFileFinderConstructor {
   pathToCheck: string;
@@ -35,7 +41,7 @@ export class DuplicateFileFinder {
 
         return {
           name: fileName,
-          path: Util.getPath(config.pathToCheck, fileName),
+          path: Util.getPath(defaultConfig.pathToCheck, fileName),
         };
       });
     });
@@ -46,7 +52,7 @@ export class DuplicateFileFinder {
   private async readFileListAsBuffers(fileList: string[]): Promise<IMapedFile[]> {
     const timerLabel = `Read ${fileList.length} files in`;
     console.time(timerLabel);
-    const promises = fileList.map((name) => fsPromise.readFile(Util.getPath(config.pathToCheck, name)));
+    const promises = fileList.map((name) => fsPromise.readFile(Util.getPath(defaultConfig.pathToCheck, name)));
     const buffers = await Promise.all(promises);
     const map = fileList.map((name, index) => ({
       name,
