+++ src/models/util.ts
